Add subscriber store
Basic in memory store to get us going. Playing around with the new
support for async traits natively.

May want to think about moving the in memory implementation elsewhere as
we grow, but seems okay for now.

Also likely want proper error handling in the future via anyhow or
similar.