Add subscriber routes
Right now these just take in the sqlite pool directly and then build
the store. We probably want to figure out a way to have them take in
something else, maybe just an impl or something that can get them a
store. I didn't want to just add the store via an Arc because we need
the store to be mutable. Which _maybe_ we could change, but for now I
just want to see some pieces working.